+++ HISTORY.rst
@@ -13,6 +13,7 @@ Unreleased
 - #787 Support .MPO image files
 - #849 expose `_Cell.row_idx` and `col_idx`
 - fix: #1042 Image.content_type and Image.ext incorrect for EMF files
+- #702 reproducible builds via fixed zip timestamps
 
 1.0.2 (2024-08-07)
 ++++++++++++++++++